ios: use a worker queue for all WebRTC operations
WebRTC does not require that we call its functions in the UI thread. We
still need to do that when using AFVoundation APIs, for example, but all
PeerConnection and MediaStream APIs can be called from a non-main thread
without a problem.

This frees the main thread from doing some unnecessary work.